+++ HISTORY.md
@@ -23,7 +23,9 @@ Our backwards-compatibility policy can be found [here](https://github.com/python
   ([#699](https://github.com/python-attrs/cattrs/issues/699))
 - _cattrs_ now tracks performance using [codspeed](https://codspeed.io/python-attrs/cattrs).
   ([#703](https://github.com/python-attrs/cattrs/pull/703))
-- The {mod}`tomlkit <cattrs.preconf.tomlkit>` preconf converter now properly handles `date` and `datetime` objects when structuring.
+- The {mod}`tomlkit <cattrs.preconf.tomlkit>` preconf converter now properly handles native `date` objects when structuring.
+  ([#707](https://github.com/python-attrs/cattrs/issues/707) [#708](https://github.com/python-attrs/cattrs/pull/708))
+- The {mod}`tomlkit <cattrs.preconf.tomlkit>` preconf converter now passes date objects directly to _tomlkit_ for unstructuring.
   ([#707](https://github.com/python-attrs/cattrs/issues/707) [#708](https://github.com/python-attrs/cattrs/pull/708))